refactor: separate selector from hook useTypedSignSimulationEnabled - to use with react class components
