+++ app/selectors/signatureController.ts
@@ -1,4 +1,8 @@
-import { SignatureRequest } from '@metamask/signature-controller';
+import { SignTypedDataVersion } from '@metamask/eth-sig-util';
+import { MessageParamsTyped, SignatureRequest, SignatureRequestType } from '@metamask/signature-controller';
+import { selectUseTransactionSimulations } from './preferencesController';
+import { isRecognizedPermit } from '../components/Views/confirmations/utils/signature';
+import { isNonPermitSupportedByDecodingAPI } from '../components/Views/confirmations/utils/signatureDecoding';
 import { RootState } from '../reducers';
 import { createDeepEqualSelector } from './util';
 
@@ -15,4 +19,31 @@ export const selectSignatureRequestById = createDeepEqualSelector(
   (_: RootState, id: string) => id,
   (signatureRequests, id) =>
     signatureRequests[id] as SignatureRequest | undefined,
+);
+
+export const selectTypedSignSimulationEnabled = createDeepEqualSelector(
+  [selectSignatureRequestById, selectUseTransactionSimulations],
+  (signatureRequest, useTransactionSimulations) => {
+    if (!signatureRequest) {
+      return undefined;
+    }
+
+    const requestType = signatureRequest.type;
+    const signatureMethod = (signatureRequest.messageParams as MessageParamsTyped)?.version;
+
+    const isTypedSignV3V4 = requestType === SignatureRequestType.TypedSign && (
+      signatureMethod === SignTypedDataVersion.V3 ||
+      signatureMethod === SignTypedDataVersion.V4
+    );
+    const isPermit = isRecognizedPermit(signatureRequest);
+
+    const nonPermitSupportedByDecodingAPI: boolean =
+      isTypedSignV3V4 && isNonPermitSupportedByDecodingAPI(signatureRequest);
+
+    return (
+      useTransactionSimulations &&
+      isTypedSignV3V4 &&
+      (isPermit || nonPermitSupportedByDecodingAPI)
+    );
+  }
 );
